pyproject.toml
[project]
name = "rag-langchain"
version = "0.1.0"
description = "Minimal LangChain RAG skeleton"
requires-python = ">=3.11"
dependencies = [
  "langchain>=1.0.0",
  "langgraph>=0.4.0",
  "langchain-openai>=1.0.0",
  "langchain-chroma>=0.1.2",
  "langchain-community>=0.3.0",
  "langchain-text-splitters>=0.3.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "pypdf>=5.0.0",
]

[tool.setuptools]
py-modules = []